+++ include/samurai/mr/adapt.hpp
@@ -14,66 +14,13 @@
 #include "../boundary.hpp"
 #include "../field.hpp"
 #include "../timers.hpp"
+#include "config.hpp"
 #include "criteria.hpp"
 #include "operators.hpp"
+#include "rel_detail.hpp"
 
 namespace samurai
 {
-    class mra_config
-    {
-      public:
-
-        auto& epsilon(double eps)
-        {
-            m_epsilon = eps;
-            return *this;
-        }
-
-        auto& epsilon() const
-        {
-            return m_epsilon;
-        }
-
-        auto& regularity(double reg)
-        {
-            m_regularity = reg;
-            return *this;
-        }
-
-        auto& regularity() const
-        {
-            return m_regularity;
-        }
-
-        auto& relative_detail(bool rel)
-        {
-            m_rel_detail = rel;
-            return *this;
-        }
-
-        auto& relative_detail() const
-        {
-            return m_rel_detail;
-        }
-
-        void init_options(CLI::App& app_)
-        {
-            app_.add_option("--mr-eps", m_epsilon, "The epsilon used by the multiresolution to adapt the mesh")
-                ->capture_default_str()
-                ->group("Multiresolution");
-            app_.add_option("--mr-reg", m_regularity, "The regularity criteria used by the multiresolution to adapt the mesh")
-                ->capture_default_str()
-                ->group("Multiresolution");
-            app_.add_flag("--mr-rel-detail", m_rel_detail, "Use relative detail instead of absolute detail")->group("Multiresolution");
-        }
-
-      private:
-
-        double m_epsilon    = 1e-4;
-        double m_regularity = 1.;
-        bool m_rel_detail   = false;
-    };
-
     struct stencil_graduation
     {
         static auto call(samurai::Dim<1>)
@@ -298,83 +245,6 @@ namespace samurai
         }
     }
 
-    namespace detail
-    {
-        void set_inv_max_field(auto& inv_max_fields, const auto& field, std::size_t dec = 0)
-            requires(std::decay_t<decltype(field)>::is_scalar)
-        {
-            auto& mesh = field.mesh();
-
-            for_each_cell(mesh,
-                          [&](const auto& cell)
-                          {
-                              inv_max_fields[dec] = std::max(inv_max_fields[dec], std::abs(field[cell]));
-                          });
-        }
-
-        void set_inv_max_field(auto& inv_max_fields, const auto& field, std::size_t dec = 0)
-            requires(!std::decay_t<decltype(field)>::is_scalar)
-        {
-            auto& mesh = field.mesh();
-
-            for_each_cell(mesh,
-                          [&](const auto& cell)
-                          {
-                              for (std::size_t i = 0; i < field.n_comp; ++i)
-                              {
-                                  inv_max_fields[dec + i] = std::max(inv_max_fields[dec + i], std::abs(field[cell][i]));
-                              }
-                          });
-        }
-
-        template <class... TFields>
-        void set_inv_max_field(auto& inv_max_fields, const Field_tuple<TFields...>& field)
-        {
-            auto f = [](auto& inv_max_fields, const auto& field, auto& dec)
-            {
-                set_inv_max_field(inv_max_fields, field, dec);
-                dec += field.n_comp;
-            };
-
-            std::size_t dec = 0;
-            std::apply(
-                [&](const auto&... args)
-                {
-                    (f(inv_max_fields, args, dec), ...);
-                },
-                field.elements());
-        }
-
-        auto compute_relative_detail(auto& detail, const auto& fields)
-        {
-            using detail_t   = typename std::decay_t<decltype(detail)>;
-            using value_t    = typename detail_t::value_type;
-            using local_type = std::array<value_t, detail_t::n_comp>;
-
-            local_type inv_max_fields;
-            inv_max_fields.fill(std::numeric_limits<value_t>::min());
-
-            set_inv_max_field(inv_max_fields, fields);
-
-#ifdef SAMURAI_WITH_MPI
-            mpi::communicator world;
-            mpi::all_reduce(world, mpi::inplace(inv_max_fields.data()), inv_max_fields.size(), mpi::maximum<value_t>());
-#endif
-
-            for (std::size_t i = 0; i < inv_max_fields.size(); ++i)
-            {
-                if (inv_max_fields[i] < std::numeric_limits<value_t>::epsilon())
-                {
-                    inv_max_fields[i] = 1.0;
-                }
-                inv_max_fields[i] = 1. / inv_max_fields[i];
-            }
-
-            auto inv_max_fields_xt = xt::adapt(inv_max_fields);
-            detail.array() *= inv_max_fields_xt;
-        }
-    } // namespace detail
-
     template <bool enlarge_, class PredictionFn, class TField, class... TFields>
     template <class... Fields>
     bool Adapt<enlarge_, PredictionFn, TField, TFields...>::harten(std::size_t ite, const mra_config& cfg, Fields&... other_fields)
@@ -463,7 +333,7 @@ namespace samurai
 
         if (cfg.relative_detail())
         {
-            detail::compute_relative_detail(m_detail, m_fields);
+            compute_relative_detail(m_detail, m_fields);
         }
 
         // std::cout << fmt::format("[{}] MR mesh adaptation step 3", world.rank()) << std::endl;
